chore(split-environments): recreate auth-with-federated environment (#2704)
* Use auth-with-federated again

* Add back auth-with-federated environment

* Back to zeroconfig!

* Skip apple test

* Fix typo
