+++ environments/auth/auth-with-federated/README.md
@@ -1,0 +1,16 @@
+# Auth with Federated
+
+This backend is configured with Amplify Admin UI:
+
+- Authentication
+  - `Email` is a required attribute, if not using a social provider to sign in
+  - `Amazon`, `Facebook`, and `Google` are all supported to sign in with
+  - No screenshots are included as they contain secret IDs for the web apps
+
+## Using this Backend
+
+External contributors can re-create this backend by running:
+
+```shell
+amplify init
+```
