+++ environments/auth/auth-with-federated/amplify/backend/types/amplify-dependent-resources-ref.d.ts
@@ -1,0 +1,15 @@
+export type AmplifyDependentResourcesAttributes = {
+    "auth": {
+        "authwithfederatedd9f06071": {
+            "IdentityPoolId": "string",
+            "IdentityPoolName": "string",
+            "HostedUIDomain": "string",
+            "OAuthMetadata": "string",
+            "UserPoolId": "string",
+            "UserPoolArn": "string",
+            "UserPoolName": "string",
+            "AppClientIDWeb": "string",
+            "AppClientID": "string"
+        }
+    }
+}
