+++ environments/auth/auth-with-federated/package.json
@@ -1,0 +1,8 @@
+{
+  "private": true,
+  "name": "auth-with-email-environment",
+  "version": "0.0.1",
+  "scripts": {
+    "pull": "amplify pull --appId d37a2ez9przpqq --envName dev"
+  }
+}
