+++ examples/angular/src/pages/ui/components/authenticator/sign-in-federated/aws-exports.js
@@ -1,2 +1,2 @@
-import awsExports from '@environments/auth/auth-with-email/src/aws-exports';
+import awsExports from '@environments/auth/auth-with-federated/src/aws-exports';
 export default awsExports;
