+++ examples/angular/src/pages/ui/components/authenticator/sign-in-federated/sign-in-federated.component.html
@@ -1,6 +1,4 @@
-<amplify-authenticator
-  [socialProviders]="['amazon', 'apple', 'facebook', 'google']"
->
+<amplify-authenticator>
   <ng-template
     amplifySlot="authenticated"
     let-user="user"
