+++ examples/vue/src/pages/ui/components/authenticator/sign-in-federated/index.vue
@@ -8,7 +8,7 @@ Amplify.configure(aws_exports);
 </script>
 
 <template>
-  <authenticator :social-providers="['amazon', 'apple', 'facebook', 'google']">
+  <authenticator>
     <template v-slot="{ user, signOut }">
       <h1>Hello {{ user.username }}!</h1>
       <button @click="signOut">Sign Out</button>
