+++ packages/e2e/features/ui/components/authenticator/sign-in-federated.feature
@@ -9,7 +9,7 @@ Feature: Sign In with Federation
   @angular @react @vue
   Scenario: Sign In screen has social providers
     Then I see "Sign In with Amazon"
-    And I see "Sign In with Apple"
+    # And I see "Sign In with Apple"
     And I see "Sign In with Facebook"
     And I see "Sign In with Google"
 
@@ -17,6 +17,6 @@ Feature: Sign In with Federation
   Scenario: Sign Up screen has social providers
     When I click the "Create Account" tab
     Then I see "Sign Up with Amazon"
-    And I see "Sign Up with Apple"
+    # And I see "Sign Up with Apple"
     And I see "Sign Up with Facebook"
     And I see "Sign Up with Google"
